+++ CHANGELOG.md
@@ -25,6 +25,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Added support for multiple architectures to `sat bootprep`, which includes the
+  following:
+    - The ability to filter the base image or recipe from a product based on the
+      architecture of the image or recipe in IMS
+    - The ability to combine multiple base image or recipe filters when
+      specifying a base image from a product
+    - The ability to specify the architecture for BOS session templates, either
+      at the top level or per boot set
+
 ## [3.23.0] - 2023-06-08
 
 ### Added